Split Figure 3 FGS CLI entry point
